Add files via upload
The first part of qs 3.1 is to apply knn algorithm to classify data, and measure the fitness of the algorithm afterwards. The second part of qs 3.1 is to find the optimal number of k.
Qs 4.2 is about to apply kmeans to classify data, using elbow method and silhouette score to find the optimal k. Occasionally, using elbow method is one of the plausible method, but it can provide more confident complement with silhouette score.